+++ src/Domain/Entities/Property.cs
@@ -10,7 +10,7 @@ public class Property
     public decimal Price { get; set; }
     public PropertyStatus Status { get; set; }
     public int Bedrooms { get; set; }
-    public decimal SurfaceArea { get; set; }
+    public int SurfaceArea { get; set; }
     public DateTimeOffset CreatedDate { get; set; }
     public int? AgencyId { get; set; }
     public Guid? UserId { get; set; }
